+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## TO BE RELEASED
 
+## v1.34.0 - 10/29/3018
+
 * OPC-149 Otherpubs config to combine Opencast pubs on pub listing
 
 ## v1.33.0 - 10/15/2018